requirements.txt
openai
selenium
Pillow
python-dotenv
